Fixes 1. After moving an object and undo the action the positions before the move and after the undo are not the same. The problem is caused by using screen resolution instead the real resolution for creating the undo command.
2. The move by key left of an object uses the offset of the y grid instead of
the x grid. Here the part of the patch which deals with this:

3. Problem of moving headers or footers in the undo.

svn path=/trunk/koffice/; revision=259076

+++ kpresenter/kprpage.h
@@ -263,7 +263,7 @@ public:
     bool chPic( KPresenterView *_view);
 
     //return command when we move object
-    KCommand *moveObject(KPresenterView *_view, int diffx,int diffy);
+    KCommand *moveObject(KPresenterView *_view, double diffx, double diffy);
     KCommand *moveObject(KPresenterView *m_view,const KoPoint &_move,bool key);
 
     KCommand *rotateObj(float _newAngle, bool addAngle=false);
